requirements/requirements_alex.txt
autograd==1.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
imageio==2.37.0
joblib==1.4.2
lazy_loader==0.4
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==1.12.1
torchvision==0.13.1
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0